[PATCH] dkms patch for SLES10 driver disk
This patch fixes the way driver disks are created for SLES10... there
are two issues that are fixed:

1)  dkms is putting the modules in an "updates" directory instead of the
default path (for example, it was using
"/lib/modules/2.6.16.21-0.8-smp/updates" instead of
"/lib/modules/2.6.16.21-0.8-smp/kernel/drivers/scsi/megaraid")... this
caused the new driver to not be used after the installer rebooted the
system because the original driver was still in the original location.
I'm not sure why it was using "updates" in the first place...

2)  dkms was not copying the driver for the default kernel into the
right modules directory on the driver disk, so SLES10 wasn't using the
new driver during installation.  There were two problems--dkms was
trying to copy the modules from the wrong directory, and also there were
quotes around the source in the copy command so the "*" wasn't being
resolved.

Patch attached.

Signed-off-by: Matt Domsch <Matt_Domsch@dell.com>