debian/patches/16_dpkg-divert.patch: Use dpkg-divert to override an existing kernel module, thanks to Andrea Mennucci. (Closes: #529059)
when dkms installs, it may override an existing kernel module
by a new one. To this end, dkms moves away the old module using 'mv'

But this breaks if the kernel is upated/reinstalled (as for example
after the recent security updates).

There is though a more effective way, and it is to use dpkg-divert