Insert the package version directly into postinst at creation time
dh_dkms already knows the package version, so it can insert it directly
into the package's postinst, instead of trying to figure it out during
the installation.  This was already the case for prerm, but not for
postinst.

(Besides, the dpkg-query invocation did not account for '-' being
allowed in the upstream version.)

This also renames $DKMS_CVERSION to $DKMS_VERSION, in line with prerm.