Remove old modules when removing a kernel
Closes: #586724
Thanks: Jan Muszynski

When removing a kernel it would be good to remove all the modules and not just uninstall them.
Issuing a remove instead of an uninstall to dkms will cause the module to be uninstalled first
(if it's not already), then removed. If the kernel is ever re-installed then dkims will
rebuild the module as needed, so nothing is lost by removing the modules for a
reomoved kernel, but you do run a cleaner system.