Add Reset-build-dependencies.patch
For each iteration we need to rebuild ithe list of uninstalled
build dependencies.  To do this next_depends must be reset for
each pass so it can be recalculated.  Otherwise, we can hit the
install_count early return and not install all the requested
packages.

The code should also use 'continue 2' instead of 'break 2' to
ensure the full list of build_depends is checked.